requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.6
pycodestyle==2.7.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
